Switch to using floats instead of integer

+++ db/migrate/20210705154113_create_counter_values.rb
@@ -2,7 +2,7 @@ class CreateCounterValues < ActiveRecord::Migration[6.1]
   def change
     create_table :counter_values do |t|
       t.string :name, index: true
-      t.integer :value, default: 0
+      t.float :value, default: 0
       t.references :parent, polymorphic: true
 
       t.timestamps
